Ensure after_sign_in_path overwrite resolves paths correctly
Our Keycloak Engine is isolated, means it cannot access resources in the main application. Before Rails 6.1, it was able to resolve the first_page_path from our dummy test app, despite it being included in the main application.

Rails 6.1 seems to fix this, as I got a ActionController::UrlGenerationError. By specifying main_app in the front of the paths, the engine will call the application it was mounted in to resolve the paths.